Use OSD count when calculating recovery window
The recovery window was previously calculated based on the number of
pgs and the recovery rate. This is problematic because the recovery
throughput could exceed the combined throughput achievable by the
OSDs. This change requests the total number of OSDs and uses that
to determine the recovery rate instead of number of PGs.